+ daemon exit fixing not finished

+++ projects/sh/abrt_server.sh
@@ -6,8 +6,11 @@ if [[ -f "sh/setup.sh" ]] ; then
   sig=$1
   shift
   if ! [[ "$sig" ]] ; then sig=QUIT ; fi
-  spid=$( server_pid )
-  if [[ "$spid" ]] ; then
+  if spid=$( server_pid ) && [[ "$spid" ]] ; then
+    echo "killing $spid with $sig" >&2
     kill -$sig $spid
+  else
+    echo "No server pid" >&2
   fi
+  psshow
 fi
